UnifiedMCP/requirements.txt
fastmcp>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-mcp-adapters>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.112.0
uvicorn>=0.30.0
httpx>=0.27.0
